lab2/requirements.txt
pika
python-dotenv
fastapi
uvicorn
